Demonstrate more complex use of locks
This commit introduces a "lock-free" lock, which allows zero-overhead locking
when something wants to execute work in parallel, and test cases showing how
we can leverage the semaphore based locking to allow a maximum number of
concurrent workers. It is possible to have more threads doing work than the
maximum number of workers by releasing a lease from a worker.